+++ test/common_widgets/key_values_dialog_test.dart
@@ -99,7 +99,8 @@ void main() {
         findsOneWidget,
       );
 
-      await safeTapByFinder(tester, find.text('Apply'));
+      await tester.drag(find.byType(DraggableScrollableSheet),
+          const Offset(0, 500)); // Dismiss dialog.
       await tester.pumpAndSettle();
 
       expect(